#68 - moved shipping address to separate component
